tools: Split "generate" container command into raw/library
