--- tools/Google.Cloud.Tools.ReleaseManager/ContainerCommands/GenerateCommand.cs
@@ -1,81 +1,0 @@
-// Copyright 2024 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License"):
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using Google.Cloud.Tools.Common;
-using System;
-using System.Linq;
-
-namespace Google.Cloud.Tools.ReleaseManager.ContainerCommands;
-
-/// <summary>
-/// Generates files for either a single API or all configured APIs. Expected options:
-/// - api-root: effectively the googleapis directory; required
-/// - generator-input: optional, when omitted, run "raw generation"
-/// - output: root folder for result; required, must exist
-/// - api-path: (relative to api-root) e.g. google/cloud/functions/v2
-//    optional when generator-input is specified, when omitted, all configured APIs should be regenerated
-//    required when generator-input is not specified (because we don’t want to raw-generate everything in api-root)
-/// </summary>
-internal class GenerateCommand : IContainerCommand
-{
-    public int Execute(ContainerOptions options)
-    {
-        string apiRoot = options.RequireOption(options.ApiRoot);
-        string output = options.RequireOption(options.Output);
-        string generatorInput = options.GeneratorInput;
-        string apiPath = options.ApiPath;
-
-        if (apiPath is null && generatorInput is null)
-        {
-            throw new UserErrorException("At least one of api-path or generator-input must be specified");
-        }
-
-        // For unconfigured generation, we use a really minimal generator-input directory in the container itself.
-        // At the time of writing, this is only used for the common resource config (which hadn't changed between March 2021
-        // and December 2024, so is clearly pretty stable).
-        var rootLayout = RootLayout.ForGeneration(generatorInput ?? "/app/generator-input", output, apiRoot);
-
-        // Note: we expect the container to already have environment variables for
-        // protoc, protobuf tools root, the gRPC generator, and the GAPIC generator.
-        var generatorCommand = new GenerateApisCommand(rootLayout);
-        if (generatorInput is null)
-        {
-            return generatorCommand.Execute(new[] { "--unconfigured", apiPath });
-        }
-        else
-        {
-            // Set environment variables used by scripts invoked by GenerateApisCommand.
-            Environment.SetEnvironmentVariable(GenerateApisCommand.GeneratorInputDirectoryEnvironmentVariable, rootLayout.GeneratorInput);
-            Environment.SetEnvironmentVariable(GenerateApisCommand.GeneratorOutputDirectoryEnvironmentVariable, rootLayout.GeneratorOutput);
-            Environment.SetEnvironmentVariable(GenerateApisCommand.GoogleApisDirectoryEnvironmentVariable, rootLayout.Googleapis);
-
-            if (apiPath is not null)
-            {
-                var catalog = ApiCatalog.Load(rootLayout);
-                var id = catalog.Apis.FirstOrDefault(api => api.ProtoPath == apiPath)?.Id;
-                if (id is null)
-                {
-                    // TODO: Optionally, create a temporary configuration.
-                    Console.WriteLine($"No configured API for path '{apiPath}'. Aborting.");
-                    return 1;
-                }
-                return generatorCommand.Execute(new[] { id });
-            }
-            else
-            {
-                return generatorCommand.Execute(new string[0]);
-            }
-        }
-    }
-}
